feat: simplify help page components by replacing existing logic with HelpView component
